requirements.txt
absl-py==2.1.0
accelerate==0.31.0
aiofiles==23.2.1
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
black==21.4b2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
constantly==23.10.4
contourpy==1.3.0
cryptography==43.0.0
cycler==0.12.1
Cython==3.0.10
cython
datasets==2.20.0
debugpy==1.8.2
decorator==5.1.1
deepspeed==0.14.4
dill==0.3.8
dnspython==2.6.1
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.0
einops-exts==0.0.4
email_validator==2.2.0
exceptiongroup==1.2.1
executing==2.0.1
fairscale==0.4.13
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.4
flash-attn==2.5.9.post1
fonttools==4.54.1
frozenlist==1.4.1
fsspec
ftfy==6.2.0
future==1.0.0
fvcore==0.1.5.post20221221
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
gradio==4.38.1
gradio_client==1.1.0
grpcio==1.65.4
h11==0.14.0
hjson==3.1.0
hkdf==0.0.3
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.26.2
humanize==4.10.0
hydra-core==1.3.2
hyperlink==21.0.0
idna==3.7
imageio==2.36.0
importlib_metadata==8.2.0
importlib_resources==6.4.0
incremental==24.7.2
iopath==0.1.9
ipykernel==6.29.5
ipython==8.26.0
iterable-io==1.0.0
jedi==0.19.1
Jinja2==3.1.4
jpeg4py==0.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
lxml==5.2.2
magic-wormhole==0.14.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
motmetrics==1.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
MultiScaleDeformableAttention==1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.1.0
ninja==1.11.1.1
numpy==2.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open-clip-torch==2.0.2
opencv-python==4.10.0.84
orjson==3.10.6
packaging==24.1
pandas==2.2.2
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@7bb4655548f98f3fedc07bf37e9040a992b054b0
parso==0.8.4
pathspec==0.12.1
peft==0.11.1
pexpect==4.9.0
pillow==10.4.0
pip-date==1.0.5
platformdirs==4.2.2
portalocker==2.10.1
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools @ git+https://github.com/XD7479/cocoapi.git@772b85a787ef554a732d4cb9abe317d61602af83#subdirectory=PythonAPI
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
pydot==3.0.1
pydub==0.25.1
Pygments==2.18.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
ruff==0.5.4
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.7.1
service-identity==24.1.0
setproctitle==1.3.3
Shapely==1.7.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spake2==0.8
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
tifffile==2024.9.20
tikzplotlib==0.10.1
timm==1.0.8
tokenizers==0.20.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
transformers==4.46.1
triton==2.3.1
Twisted==24.3.0
txaio==23.1.1
txtorcon==23.11.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
visdom==0.2.4
wandb==0.17.3
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==24.6.0
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.3
wormhole==0.1.0
xmltodict==0.13.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
zipp==3.19.2
zipstream-ng==1.7.1
zope.interface==6.4.post2

